+++ sinch/domains/numbers/api/v1/__init__.py
@@ -1,11 +1,11 @@
 from sinch.domains.numbers.api.v1.active_numbers_apis import ActiveNumbers
 from sinch.domains.numbers.api.v1.available_numbers_apis import AvailableNumbers
 from sinch.domains.numbers.api.v1.available_regions_apis import AvailableRegions
-from sinch.domains.numbers.api.v1.numbers_callbacks_apis import Callback
+from sinch.domains.numbers.api.v1.callback_configuration_apis import CallbackConfiguration
 
 __all__ = [
     "ActiveNumbers",
     "AvailableNumbers",
     "AvailableRegions",
-    "Callback"
+    "CallbackConfiguration"
 ]
