+++ sinch/domains/numbers/api/v1/internal/__init__.py
@@ -6,7 +6,7 @@ from sinch.domains.numbers.api.v1.internal.available_numbers_endpoints import (
     ActivateNumberEndpoint, AvailableNumbersEndpoint, RentAnyNumberEndpoint, SearchForNumberEndpoint
 )
 from sinch.domains.numbers.api.v1.internal.available_regions_endpoints import ListAvailableRegionsEndpoint
-from sinch.domains.numbers.api.v1.internal.numbers_callback_endpoints import (
+from sinch.domains.numbers.api.v1.internal.callback_configuration_endpoints import (
     GetCallbackConfigurationEndpoint, UpdateCallbackConfigurationEndpoint
 )
